requirements.txt
fastapi
uvicorn
pyzmq
redis
dill
requests
matplotlib
numpy
pytest
pydantic
psutil
pytest-asyncio
pandas
seaborn
scipy